+++ src/components/atoms/ObserveInView.tsx
@@ -7,8 +7,9 @@ import { useInView } from "react-intersection-observer";
 
 import * as styles from "./observeInView.module.css";
 
-interface ObserveInViewProps {
+interface ObserveInViewProps extends React.HTMLAttributes<HTMLDivElement> {
   animate: boolean;
+  children: React.ReactNode;
   className?: string;
   dynamicClasses?: {
     animate?: string;
@@ -22,6 +23,7 @@ interface ObserveInViewProps {
 
 export default function ObserveInView({
   animate,
+  children,
   className,
   dynamicClasses,
   observeOptions,
@@ -58,6 +60,8 @@ export default function ObserveInView({
   }, [inView, onViewChange]);
 
   return (
-    <div className={clsx(rootClasses, className)} ref={rootRef} {...props} />
+    <div className={clsx(rootClasses, className)} ref={rootRef} {...props}>
+      {children}
+    </div>
   );
 }